requirements.txt
fastapi
uvicorn
supabase
python-dotenv
influxdb
requests
pandas
openai
openai>=1.0.0